requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
certifi==2025.4.26
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
ecdsa==0.19.1
fastapi==0.115.12
fastapi-cli==0.0.7
greenlet==3.2.1
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
packaging==25.0
passlib==1.7.4
pluggy==1.5.0
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
pytest==8.3.5
pytest-asyncio==0.26.0
python-dotenv==1.1.0
python-jose==3.4.0
PyYAML==6.0.2
rich==14.0.0
rich-toolkit==0.14.3
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
watchfiles==1.0.5
websockets==15.0.1
